Yet more nullable fields in the wild Yet more fields need to be nullable More fields need to be nullable Fixed enumerations so their numeric values do NOT change for C# binary API compatibility. Add in missing error response This must also be nullable These are NULLABLE, as they can come back as NULL from the addressd validation service. I am not sure why you guys changed then to not allow null??? Add support for missing label_image_id which is not documented on the web site. Added in the missing carrier information to the error responses Some countries do not have states, so this can fail if null is passed for an empty country Gah, this *also* needs to be nullable This must be nullable, as it comes back null for USPS manifest documents from Endicia. Fixed incorrectly specified 204 responses as they are actually empty and there is no response content to parse These should not be marked as obsolete as it will cause the compiler to throw and error if you use it. And it still does work. Tracking numbers need to be optional to correctly parse the response when a label is a test label with USPS. Warehouse ID is *not* required for manifesting, and this will fail to parse the responses sent back if you manifest for specific labels. Add support for undocumented test_labels parameters Actually there is already a type for carrier code The carrier code is a general string, NOT an se_id. There is an issue with the spec I am not sure how to solve. The first simpler problem is that customs_item_id is marked as required and not-nullable, but it really should be nullable so it can be left out of the request. However the real issue is with the value property, as it's defined as:
value:
  allOf:
    - $ref: '#/components/schemas/monetary_value'
  description: The total value of the customs item

and a monetary value is expected to be a structure that contains the amount and currency type, so it's passed in the request like so:

Value = new MonetaryValue {
    Amount = 42.99,
    Currency = Currency.Usd,
},

but the problem is, in the response, it comes back as a single numeric value:

    "customs_items": [
      {
        "customs_item_id": "se-11825779",
        "description": "Product name",
        "quantity": 2,
        "value": 42.99,
        "harmonized_tariff_code": "",
        "country_of_origin": "US",
        "unit_of_measure": null
      }
    ],

So that is completely incorrect as far as the OpenAPI spec is concerned, but it's not clear to me how to fix it?

For now the service seems fine with just passing a raw number and it assumes it is in USD, so I am going with that for now.
These can be null for UPS SurePost, so the spec needs to handle that
These all need to be nullable also as the service responses actually have nulls in them
This item is nullable as the actual code can return null here (included in JSON, but null). So either we need to mark it as nullable here or the service should not be including the value.
Fixed mailer-id to be common type, so it works with C# client generation